+++ artofinvesting-frontend/package.json
@@ -9,6 +9,7 @@
     "antd": "^4.4.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1"
   },
   "scripts": {
