--- artofinvesting-frontend/src/NavBar.js
@@ -1,29 +1,0 @@
-import React, { Component } from 'react'
-import { Tabs } from 'antd';
-const { TabPane } = Tabs;
-
-
-export default class NavBar extends Component {
-
-    callBack = (key) => {
-        console.log(key);
-    }
-
-    render() {
-        return (
-            <div>
-                <Tabs defaultActiveKey="1" onChange={this.callback}>
-                    <TabPane tab="Tab 1" key="1">
-                        Content of Tab Pane 1
-                    </TabPane>
-                    <TabPane tab="Tab 2" key="2">
-                        Content of Tab Pane 2
-                    </TabPane>
-                    <TabPane tab="Tab 3" key="3">
-                        Content of Tab Pane 3
-                    </TabPane>
-                </Tabs>
-            </div>
-        )
-    }
-}
