+++ artofinvesting-frontend/src/components/CompanySearchBar.js
