+++ artofinvesting-frontend/src/components/HomePage.js
@@ -1,37 +1,13 @@
 import React, { Component } from 'react';
-import { Button, Row, Col, Modal } from 'antd';
-import logo from './aoi_logo.png';
-import LoginForm from './LoginForm.js';
+import { Button, Row, Col } from 'antd';
+import logo from '../aoi_logo.png';
 import CompanySearchBar from './CompanySearchBar';
+import SignInButton from './SignInButton';
 
 class HomePage extends Component {
 
     constructor(props) {
         super(props);
-        this.state = {
-            isSignInPopUpVisible: false
-        };
-
-    }
-
-    onSignInButtonClick = () => {
-        this.setState({
-            isSignInPopUpVisible: true
-        });
-    }
-
-    handleOk = (e) => {
-        console.log(e);
-        this.setState({
-            isSignInPopUpVisible: false,
-        });
-    }
-
-    handleCancel = (e) => {
-        console.log(e);
-        this.setState({
-            isSignInPopUpVisible: false,
-        }); 
     }
 
 
@@ -40,16 +16,9 @@ class HomePage extends Component {
             <div>
                 <Row style={{ "paddingTop": "20px" }} type="flex" justify="end" align="bottom">
                     <Col pull={1}>
-                        <Button type="primary" onClick={this.onSignInButtonClick}>Sign in</Button>
+                        <SignInButton/>
                     </Col>
                 </Row>
-                <Modal 
-                    visible={this.state.isSignInPopUpVisible}
-                    onOk={this.handleOk}
-                    onCancel={this.handleCancel}
-                >
-                    <LoginForm />
-                </Modal>
                 <Row justify="center" align="top" style={{ "paddingTop": "100px" }}>
                     <img src={logo} className="AOI-logo" alt="logo" />
                 </Row>
